added test using bigger databases; added small benchmarking application; added overflow counter to the search results

+++ src/libssa.c
@@ -43,7 +43,7 @@ void set_verbose( int debug ) {
     _verbose = debug;
 }
 
-void set_threads( long nr ) {
+void set_threads( size_t nr ) {
     _max_thread_count = nr;
 }
